requirements.txt
# general  modules
django >3.2,<3.3
Pillow
djangorestframework

# env controlls
python-decouple

# third party modules
dj-rest-auth[with_social]
markdown
django-filter  
drf-yasg[validation]
djangorestframework-simplejwt
django-cors-headers
django-csp
django-robots
django-debug-toolbar

# deployment module
gunicorn

# refactoring and tests
black
flake8
pytest
pytest-django
Faker

# background proccess and cache
celery
django-redis

# database client
psycopg2-binary

